milestone_2/week_3/Day_11/demo/requirements.txt
huggingface_hub
python-dotenv
tiktoken